CSV to GDF script
Python script created to process .csv file generated by
get_rugby_data.py and to create the necessary node and edge definitions
for the GDF file.

A node is considered as each unique team in the data. Total score and
total tries attributes have been calculated for each node

An edge is considered as a fixture between two teams. The weighting of
the edge is based on the number of times the teams have met.

Example datasets and Gephi outputs will follow